Add account-create-modal to notify on successful account creation
