+++ frontend/src/app/components/modals/account-created-modal/account-created-modal.html
@@ -1,0 +1,12 @@
+<p class="highlight push-next-half">
+    Copy and send the following information to the account owner or connect an application:
+</p>
+
+<div class="alt-bg pad row content-start greedy push-next">
+    <pre class="greedy message" ko.html="message"></pre>
+    <copy-to-clipboard-button params="value: message"></copy-to-clipboard-button>
+</div>
+
+<button class="btn align-end" ko.click="onClose">
+    Done
+</button>
