+++ frontend/src/app/components/modals/account-created-modal/account-created-modal.less
@@ -1,0 +1,8 @@
+.account-created-modal {
+    display: block;
+
+    .message{
+        font-family: roboto;
+        line-height: 1.5;
+    }
+}
