+++ frontend/src/app/components/modals/account-created-modal/account-details-message.html
@@ -1,11 +1,11 @@
 Hi, I created a NooBaa user for you:
 
 <span class="highlight">Use the following credentials to connect to the NooBaa console:</span>
-Console Url: <span data-bind="text: serverAddress"></span>
-Username: <span data-bind="text: username"></span>
-Password: <span data-bind="text: password"></span>
+<span>Console Url: {{serverAddress}}</span>
+<span>Username: {{username}}</span>
+<span>Password: {{password}}</span>
 <!-- ko if: accessKey && secretKey -->
 <span class="highlight">Use the following S3 access to connect an S3 compatible application to NooBaa:</span>
-Access Key: <span data-bind="text: accessKey"></span>
-Secret Key: <span data-bind="text: secretKey"></span>
+<span>Access Key: {{accessKey}}</span>
+<span>Secret Key: {{secretKey}}</span>
 <!-- /ko -->
