--- frontend/src/app/components/modals/create-account-modal/backup.html
@@ -1,71 +1,0 @@
-<editor params="label: 'Email Address'">
-    <input type="text" ko.value="bindToFormField('email')" />
-</editor>
-
-<hr />
-
-<h2 class="heading3">
-    Allow S3 access for this account
-</h2>
-
-<toggle-switch class="push-next" params="
-    value: bindToFormField('enableS3Access'),
-    onLabel: 'Enabled',
-    offLabel: 'Disabled'
-"></toggle-switch>
-
-<div class="row content-middle">
-    <h2 class="heading3 greedy push-next" ko.css.disabled="!enableS3Access()">
-        Grant S3 access permissions for buckets
-    </h2>
-    <div>
-        <button class="link alt-colors"
-            ko.click="onSelectAllBuckets"
-            ko.enable="enableS3Access"
-        >
-            Select all
-        </button>
-        <span class="disabled">|</span>
-        <button class="link alt-colors"
-            ko.click="onClearAllBuckets"
-            ko.enable="enableS3Access"
-        >
-            Clear
-        </button>
-    </div>
-</div>
-
-<multiselect class="greedy" params="
-    options: buckets,
-    selected: bindToFormField('selectedBuckets'),
-    disabled: !enableS3Access()
-"></multiselect>
-
-
-<hr />
-
-<h2 class="heading3">
-    Choose default placement for S3 client applications
-</h2>
-<editor params="label: 'Resource'">
-    <dropdown params="
-        options: resources,
-        selected: bindToFormField('selectedResource')
-    "></dropdown>
-</editor>
-
-<div class="row align-end">
-    <button class="link alt-colors push-next" data-bind="click: () => close()">
-        Cancel
-    </button>
-
-    <working-button params="
-        click: onCreate,
-        working: working,
-        workingLabel: 'Creating...'
-    ">
-        Create Account
-    </working-button>
-</div>
-
-
