+++ frontend/src/app/components/modals/create-account-modal/create-account-modal.less
@@ -1,11 +1,2 @@
 .create-account-modal {
-    hr {
-        margin-left: -@gutter;
-        margin-right: -@gutter;
-    }
-
-    .user-message{
-        font-family: roboto;
-        line-height: 1.5;
-    }
 }
