--- frontend/src/app/components/modals/create-account-modal/create-account-modal2.html
@@ -1,70 +1,0 @@
-<editor params="label: 'Email Address'">
-    <input type="text" ko.value="bindToField('email')" ko.disable="lock" />
-    <val-message params="field: email"></val-message>
-</editor>
-
-<editor params="label: 'S3 Access'">
-    <toggle-switch params="
-        value: bindToField('enableS3Access'),
-        onLabel: 'Enabled',
-        offLabel: 'Disabled',
-        disabled: lock
-    "></toggle-switch>
-</editor>
-
-<editor params="label: 'S3 default placement', disabled: !enableS3Access()">
-    <dropdown params="
-        options: resources,
-        selected: bindToField('selectedResource'),
-        disabled: !enableS3Access() || lock(),
-        placeholder: 'Choose pool...'
-    "></dropdown>
-     <val-message params="field: selectedResource"></val-message>
-</editor>
-
-<div class="row content-middle push-prev push-next-half">
-    <p class="greedy push-next" ko.css.disabled="!enableS3Access()">
-        S3 access permissions for existing buckets:
-    </p>
-    <div>
-        <button class="link alt-colors"
-            ko.click="onSelectAllBuckets"
-            ko.disable="!enableS3Access() || lock()"
-        >
-            Select all
-        </button>
-        <span class="disabled">|</span>
-        <button class="link alt-colors"
-            ko.click="onClearAllBuckets"
-            ko.disable="!enableS3Access() || lock()"
-        >
-            Clear
-        </button>
-    </div>
-</div>
-
-<multiselect class="greedy push-next" params="
-    options: buckets,
-    selected: bindToField('selectedBuckets'),
-    disabled: !enableS3Access() || lock(),
-"></multiselect>
-
-
-<div class="row align-end">
-    <button class="link alt-colors push-next"
-        ko.click="onClose"
-        ko.disable="lock"
-    >
-        Cancel
-    </button>
-
-    <working-button params="
-        click: onCreate,
-        working: lock,
-        workingLabel: 'Creating...'
-    ">
-        Create Account
-    </working-button>
-</div>
-
-
