--- frontend/src/app/components/modals/create-account-modal/create-screen.html
@@ -1,61 +1,0 @@
-<h2 class="heading3">Account Details</h2>
-
-<editor params="label: 'User Email'">
-    <input type="text" data-bind="value: emailAddress"/>
-</editor>
-
-
-<h2 class="heading3">
-    S3 access for this account
-</h2>
-
-<toggle-switch class="push-next" params="
-    value: enableS3Access,
-    onLabel: 'Enabled',
-    offLabel: 'Disabled'
-"></toggle-switch>
-
-<div class="row content-middle">
-    <h2 class="heading3 greedy push-next" data-bind="
-        css: { disabled: !enableS3Access() }
-    ">
-        Grant S3 access permissions for buckets
-    </h2>
-
-    <div>
-        <button class="link alt-colors" data-bind="
-            click: selectAllBuckets,
-            enable: enableS3Access
-        ">
-            Select all
-        </button>
-        <span class="disabled">|</span>
-        <button class="link alt-colors" data-bind="
-            click: clearAllBuckets,
-            enable: enableS3Access
-        ">
-            Clear
-        </button>
-    </div>
-</div>
-
-<multiselect class="greedy push-next" params="
-    options: buckets,
-    selected: selectedBuckets,
-    disabled: !enableS3Access()">
-</multiselect>
-
-<div class="row align-end">
-    <button class="link alt-colors push-next" data-bind="click: () => close()">
-        Cancel
-    </button>
-
-    <working-button params="
-        click: create,
-        working: working,
-        workingLabel: 'Creating...'
-    ">
-        Create Account
-    </working-button>
-</div>
-
