--- frontend/src/app/components/modals/create-account-modal/review-screen.html
@@ -1,15 +1,0 @@
-<p class="push-next">
-    <span data-bind="text: emailAddress"></span>
-    was created successfully. Accounts can be used for connecting applications to NooBaa or as admin users.
-</p>
-
-<h2 class="heading3">Copy and send the following information to the new user:</h2>
-
-<div class="alt-bg pad row content-start greedy push-next">
-    <pre class="greedy user-message" data-bind="html: userMessage"></pre>
-    <copy-to-clipboard-button params="value: userMessage"></copy-to-clipboard-button>
-</div>
-
-<button class="btn align-end" data-bind="click: close">
-    Done
-</button>
